create useTvl hook and lend query structure
