+++ apps/lend/src/entities/chain/index.ts
@@ -1,0 +1,3 @@
+export * from './lib'
+export * from './model'
+export * from './types'
