+++ apps/lend/src/entities/chain/model/chain-validation.ts
@@ -1,0 +1,12 @@
+import type { ChainQueryParams } from '@/entities/chain/types'
+import { createValidationSuite } from '@/shared/curve-lib'
+import { enforce, group, test } from 'vest'
+
+export const chainValidationGroup = ({ chainId }: ChainQueryParams) =>
+  group('chainValidation', () => {
+    test('chainId', () => {
+      enforce(chainId).message('Chain ID is required').isNotEmpty().message('Invalid chain ID').isValidChainId()
+    })
+  })
+
+export const chainValidationSuite = createValidationSuite(chainValidationGroup)
