+++ apps/lend/src/entities/chain/model/query-keys.ts
@@ -1,0 +1,5 @@
+import type { ChainQueryParams } from '@/entities/chain/types'
+
+export const chainKeys = {
+  root: ({ chainId }: ChainQueryParams) => ['chain', chainId] as const,
+} as const
