+++ apps/lend/src/entities/chain/types.ts
@@ -1,0 +1,3 @@
+export type ChainQueryParams = {
+  chainId?: ChainId
+}
